// frontend/src/components/filament-profiles/types.ts
export interface PresetForm {
  color: string;
  inherits: string;
  filament_vendor: string;
  filament_type: string;
  default_filament_colour: string;
  filament_notes: string;
  nozzle_temperature: string;
  nozzle_temperature_initial_layer: string;
  nozzle_temperature_range_low: string;
  nozzle_temperature_range_high: string;
  cool_plate_temp: string;
  cool_plate_temp_initial_layer: string;
  eng_plate_temp: string;
  eng_plate_temp_initial_layer: string;
  hot_plate_temp: string;
  hot_plate_temp_initial_layer: string;
  textured_plate_temp: string;
  textured_plate_temp_initial_layer: string;
  supertack_plate_temp: string;
  supertack_plate_temp_initial_layer: string;
  fan_max_speed: string;
  fan_min_speed: string;
  close_fan_the_first_x_layers: string;
  close_additional_fan_first_x_layers: string;
  fan_cooling_layer_time: string;
  overhang_fan_speed: string;
  during_print_exhaust_fan_speed: string;
  complete_print_exhaust_fan_speed: string;
  enable_pressure_advance: boolean;
  pressure_advance: string;
  filament_flow_ratio: string;
  filament_max_volumetric_speed: string;
  filament_prime_volume: string;
  filament_retraction_length: string;
  filament_retraction_speed: string;
  filament_retract_when_changing_layer: string;
  filament_wipe: string;
  filament_z_hop: string;
  filament_z_hop_types: string;
  filament_deretraction_speed: string;
  filament_wipe_distance: string;
  filament_retract_before_wipe: string;
  slow_down_layer_time: string;
  slow_down_min_speed: string;
  filament_cost: string;
  filament_density: string;
  filament_shrink: string;
  temperature_vitrification: string;
  chamber_temperatures: string;
  additional_cooling_fan_speed: string;
  enable_overhang_bridge_fan: string;
  first_x_layer_fan_speed: string;
  pre_start_fan_time: string;
  overhang_fan_threshold: string;
  filament_extruder_variant: string[];
  compatible_printers: string;
  filament_start_gcode: string;
  filament_end_gcode: string;
  pa_k_value: string;
  nozzle_size: string;
}

export type GridSize = 'small' | 'medium' | 'large';
export type SortField = 'name' | 'brand' | 'material' | 'color';